reoraginzed stuff into a pitch file
